REFACTOR: Move income/expenses tabs into own components
Addresses: #120
