+++ app/app/View/Components/MoneySheet/Expenses/MoneySheetExpenses.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\View\Components\MoneySheet\Expenses;
+
+use App\Livewire\Dto\MoneySheet as MoneySheetDto;
+use Domain\CoreGameLogic\EventStore\GameEvents;
+use Domain\CoreGameLogic\Feature\Moneysheet\State\MoneySheetState;
+use Domain\CoreGameLogic\Feature\Spielzug\State\PlayerState;
+use Domain\CoreGameLogic\PlayerId;
+use Illuminate\View\Component;
+use Illuminate\View\View;
+
+class MoneySheetExpenses extends Component
+{
+    /**
+     * Create the component instance.
+     */
+    public function __construct(
+        public PlayerId $playerId,
+        public GameEvents $gameStream,
+    ) {}
+
+    /**
+     * Get the view / contents that represent the component.
+     */
+    public function render(): View
+    {
+        return view('components.gameboard.moneySheet.expenses.money-sheet-expenses', [
+            'moneySheet' => new MoneySheetDto(
+                lebenskosten: MoneySheetState::calculateLebenshaltungskostenForPlayer($this->gameStream, $this->playerId),
+                steuernUndAbgaben: MoneySheetState::calculateSteuernUndAbgabenForPlayer($this->gameStream, $this->playerId),
+                gehalt: PlayerState::getGehaltForPlayer($this->gameStream, $this->playerId),
+            ),
+        ]);
+    }
+
+}
