+++ app/resources/views/components/gameboard/moneySheet/expenses/money-sheet-credits.blade.php
@@ -1,0 +1,28 @@
+@props([
+])
+
+<h3>Kredite</h3>
+<table>
+    <thead>
+    <tr>
+        <th>Kreditverwendung</th>
+        <th>Kredithöhe</th>
+        <th>Rückzahlungssumme</th>
+        <th>Rückzahlung pro Runde</th>
+        <th>offene Raten</th>
+    </tr>
+    </thead>
+    <tbody>
+    <tr>
+        <td>Kredit Phase 1</td>
+        <td>30.000 €</td>
+        <td>12.000 €</td>
+        <td>30.750 €</td>
+        <td>1.537,50 €</td>
+    </tr>
+    <tr>
+        <td colspan="4" class="text-align--right">Einnahmen Aktien gesamt</td>
+        <td>0€</td>
+    </tr>
+    </tbody>
+</table>
