+++ app/resources/views/components/gameboard/moneySheet/expenses/money-sheet-insurances.blade.php
@@ -1,0 +1,29 @@
+@props([
+])
+
+<h3>Versicherungen</h3>
+<form>
+    <div class="form__group">
+        <label>
+            <input type="checkbox">
+            100€ /Jahr Haftpflichtversicherung
+        </label>
+    </div>
+    <div class="form__group">
+        <label>
+            <input type="checkbox">
+            150€ /Jahr Private Unfallversicherung
+        </label>
+    </div>
+    <div class="form__group">
+        <label>
+            <input type="checkbox">
+            Berufsunfähigkeitsversicherung (BU)
+        </label>
+        <small>
+            Phase I: 500€/Jahr <br />
+            Einstieg in Phase II: 700€/Jahr <br />
+            Einstieg in Phase III: 900€ /Jahr
+        </small>
+    </div>
+</form>
