+++ app/resources/views/components/gameboard/moneySheet/expenses/money-sheet-kids.blade.php
@@ -1,0 +1,23 @@
+@props([
+    '$gehalt' => 0,
+])
+
+<h3>Kinder</h3>
+<table>
+    <tbody>
+    <tr>
+        <td><small>Dein Brutto gehalt</small> <br /> {{ $gehalt }} €</td>
+        <td>
+            1.400 € <br />
+            <small>
+                Pro Jahr gibst Du 5% Deines Gehaltes pro Kind aus, jedoch mindestens 1.000 €.
+            </small>
+        </td>
+        <td><small>Anzahl Kinder</small> <br />0</td>
+    </tr>
+    <tr>
+        <td colspan="2" class="text-align--right">Gesamtsumme</td>
+        <td>0 €</td>
+    </tr>
+    </tbody>
+</table>
