+++ app/resources/views/components/gameboard/moneySheet/income/money-sheet-investments.blade.php
@@ -1,0 +1,28 @@
+@props([
+])
+
+<h3>Finanzen und Vermögenswerte</h3>
+<table>
+    <thead>
+    <tr>
+        <th>Menge</th>
+        <th>Beschreibung</th>
+        <th>Kaufpreis/Stück</th>
+        <th>Dividende oder Mietertrag/Stück</th>
+        <th>Einnahmen</th>
+    </tr>
+    </thead>
+    <tbody>
+    <tr>
+        <td>1.000</td>
+        <td>Aktie XYZ</td>
+        <td>12.000 €</td>
+        <td>0 €</td>
+        <td>0 €</td>
+    </tr>
+    <tr>
+        <td colspan="4" class="text-align--right">Einnahmen Aktien gesamt</td>
+        <td>0 €</td>
+    </tr>
+    </tbody>
+</table>
