+++ app/resources/views/components/gameboard/moneySheet/money-sheet-expenses-modal.blade.php
@@ -1,0 +1,17 @@
+@props([
+    '$gameStream' => null,
+    '$playerId' => null,
+])
+
+@extends ('components.modal.modal', ['closeModal' => "toggleEditExpenses()", 'size' => 'large'])
+@section('title')
+    Money Sheet - Ausgaben
+@endsection
+
+@section('content')
+    <x-money-sheet.expenses.money-sheet-expenses :game-stream="$gameStream" :player-id="$playerId" />
+@endsection
+
+@section('footer')
+    <button type="button" class="button button--type-primary" wire:click="toggleEditExpenses()">Schließen</button>
+@endsection
